Integrate Mailhog
run `npm run dev:full-reset` to get it up and running.

This required some changes to the docker container to also contain dev-only dependencies, which then lead to a cascase of updates and eventually some hacks in the container needed to have the right ldd linked. Long story short: it works now.